requirements.txt
# GUI framework
customtkinter

# File system forensics
pytsk3

# System and disk information
psutil

# For preview functionality (images & PDFs)
Pillow
PyPDF2

# Partition recovery (if you decide to use it)
pyparted

# Packaging tool for creating installers
pyinstaller

# Run this command to install all dependencies:  pip install -r requirements.txt
